merge all feature

+++ services/product_service.py
@@ -26,7 +26,7 @@ async def get_product_byID(product_id: int):
     except SQLAlchemyError as e:
         return JSONResponse(content={"error": str(e)}, status_code=500)
 
-async def create_product(product_data: ProductSchema):
+async def create_product(product_data: ProductSchema, current_user: dict):
     try:
         product = Product(
             name=product_data.name,
@@ -34,7 +34,8 @@ async def create_product(product_data: ProductSchema):
             category=product_data.category,
             unit=product_data.unit,
             is_package=product_data.is_package,
-            image=product_data.image
+            image=product_data.image,
+            user_id=current_user["id"]  
         )
         db.session.add(product)
         db.session.commit()
@@ -42,31 +43,39 @@ async def create_product(product_data: ProductSchema):
     except SQLAlchemyError as e:
         return JSONResponse(content={"error": str(e)}, status_code=500)
     
-async def update_product(product_id: int, product_data: UpdateProductSchema):
+async def update_product(product_id: int, product_data: UpdateProductSchema, current_user: dict):
     try:
         product = db.session.query(Product).filter_by(id=product_id).first()
 
         if not product:
             return JSONResponse(content={"error": "Product not found"}, status_code=404)
         
-        update_product = product_data.dict(exclude_unset=True)
-        for key, value in update_product.items():
+        if current_user["role"] != "admin" and product.user_id != current_user["id"]:
+            return JSONResponse(content={"error": "Permission denied"}, status_code=403)
+
+        update_fields = product_data.dict(exclude_unset=True)
+        for key, value in update_fields.items():
             setattr(product, key, value)
 
         db.session.commit()
         return JSONResponse(content=product.to_dict(), status_code=200)
+
     except SQLAlchemyError as e:
         return JSONResponse(content={"error": str(e)}, status_code=500)
 
-async def delete_product(product_id: int):
+async def delete_product(product_id: int, current_user: dict):
     try:
         product = db.session.query(Product).filter_by(id=product_id).first()
 
         if not product:
             return JSONResponse(content={"error": "Product not found"}, status_code=404)
-        
+
+        if current_user["role"] != "admin" and product.user_id != current_user["id"]:
+            return JSONResponse(content={"error": "Permission denied"}, status_code=403)
+
         db.session.delete(product)
         db.session.commit()
         return JSONResponse(content={"message": "Product deleted successfully"}, status_code=200)
+
     except SQLAlchemyError as e:
         return JSONResponse(content={"error": str(e)}, status_code=500)
